requirements.txt
anthropic==0.25.6
dill==0.3.7
distlib @ file:///private/tmp/python-distlib-20231212-4938-trhlvy/distlib-0.3.8
distro==1.9.0
GitPython==3.1.32
grpcio==1.60.0
grpcio-health-checking==1.60.0
grpcio-tools==1.60.0
gymnasium==0.29.1
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.4
hyperlink==21.0.0
identify==2.5.33
idna==3.4
importlib-metadata==7.0.1
incremental==22.10.0
inquirer==3.2.1
ipykernel==6.28.0
ipython==8.20.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.1
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.0.154
langsmith==0.0.56
litellm==1.17.4
-e git+https://github.com/meta-llama/llama-stack.git@915672f86df9e9dc2c1d785bf08e5d1a16519cc8#egg=llama_stack
llama_models==0.0.42
lockfile==0.12.2
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
-e git+https://github.com/karpathy/minGPT.git@37baab71b9abea1b76ab957409a1cc2fbfba8a26#egg=minGPT
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
Naked==0.1.32
nest-asyncio==1.5.8
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.24.2
open-interpreter==0.2.0
openai==1.7.2
openapi-schema-pydantic==1.2.4
p2pnetwork==1.2
packaging==23.2
pandas==2.1.1
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
pdfplumber==0.10.2
peft==0.6.2
pexpect==4.9.0
pillow==10.3.0
platformdirs @ file:///private/tmp/python-platformdirs-20231204-5317-3q4okz/platformdirs-4.1.0
pluggy==1.3.0
posthog==3.3.1
pre-commit==3.6.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.0
PyDispatcher==2.0.7
pyflakes==2.1.1
Pygments==2.16.1
PyJWT==2.8.0
pylev==1.4.0
Pympler==1.0.1
pyparsing==3.0.5
pypdf==3.17.0
PyPDF2==3.0.1
pypdfium2==4.22.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-jobspy==1.1.25
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
python-socketio==5.11.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
queuelib==1.6.2
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==0.9.1
rich==13.5.2
rich-argparse==1.4.0
rpds-py==0.9.2
rply==0.7.8
runs==1.2.0
safetensors==0.4.1
scipy==1.11.4
Scrapy==2.11.0
semver==3.0.2
service-identity==23.1.0
sgmllib3k==1.0.0
shellescape==3.8.1
shellingham==1.5.4
shtab==1.6.5
simple-websocket==1.0.0
simple_parsing==0.1.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.18.1
swebench==1.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.5.0
textblob==0.17.1
tiktoken==0.4.0
tinysegmenter==0.3
tldextract==5.1.1
tls-client==0.2.2
together==1.1.2
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tomli==2.0.1
tomlkit==0.7.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.35.2
trl==0.7.4
Twisted==22.10.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tyro==0.6.0
tzdata==2023.3
tzlocal==4.3.1
ujson==1.35
unidiff==0.7.5
urllib3==1.26.7
validators==0.22.0
virtualenv==20.25.0
w3lib==2.1.2
watchdog==3.0.0
wcwidth==0.2.13
weaviate-client==4.4b6
websocket-client==1.7.0
Werkzeug==3.0.2
wget==3.2
wsproto==1.2.0
xmod==1.8.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.4.1
zope.interface==6.1
